app/requirements.txt
requests==2.25.1
Flask==1.1.2
gunicorn==20.0.4
google-cloud-bigquery==2.10.0
google-cloud-storage==1.36.1